DatabaseV2.php to Database.php, AbstractModel::getTable() return table without prefix
