+++ src/Builders/AbstractWithMetaBuilder.php
@@ -14,7 +14,6 @@ use Dbout\WpOrm\Exceptions\WpOrmException;
 use Dbout\WpOrm\MetaMappingConfig;
 use Dbout\WpOrm\Orm\AbstractModel;
 use Dbout\WpOrm\Orm\Database;
-use Dbout\WpOrm\Orm\DatabaseV2;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -130,7 +129,7 @@ abstract class AbstractWithMetaBuilder extends AbstractBuilder
             $join->on(
                 sprintf('%s.%s', $metaKey, $this->metaConfig?->columnKey),
                 '=',
-                DatabaseV2::getInstance()->raw(sprintf("'%s'", $metaKey))
+                Database::getInstance()->raw(sprintf("'%s'", $metaKey))
             )->on(
                 sprintf('%s.%s', $metaKey, $this->metaConfig?->foreignKey),
                 '=',
