+++ src/Orm/AbstractModel.php
@@ -8,8 +8,8 @@
 
 namespace Dbout\WpOrm\Orm;
 
+use Illuminate\Database\Connection;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Str;
 
 abstract class AbstractModel extends Model
 {
@@ -30,7 +30,7 @@ abstract class AbstractModel extends Model
     /**
      * @inheritDoc
      */
-    protected function newBaseQueryBuilder()
+    protected function newBaseQueryBuilder(): Builder
     {
         $connection = $this->getConnection();
         return new Builder(
@@ -43,28 +43,9 @@ abstract class AbstractModel extends Model
     /**
      * @inheritDoc
      */
-    public function getConnection()
+    public function getConnection(): Connection
     {
-        // @phpstan-ignore-next-line
-        return DatabaseV2::getInstance();
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function getTable(): string
-    {
-        $prefix = $this->getConnection()->getTablePrefix();
-
-        if (!empty($this->table)) {
-            return str_starts_with($this->table, $prefix) ? $this->table : $prefix . $this->table;
-        }
-
-        $table = \substr(\strrchr(\get_class($this), "\\"), 1);
-        $table = Str::snake(Str::plural($table));
-
-        // Add WordPress table prefix
-        return $prefix . $table;
+        return Database::getInstance();
     }
 
     /**
